Update and add more tests
