+++ tests/unit/dataactvalidator/validation_handlers/test_file_generation_manager.py
@@ -1,184 +1,238 @@
-from dataactcore.models.jobModels import FileType, JobStatus, JobType
-from dataactvalidator.validation_handlers import file_generation_manager
-from tests.unit.dataactcore.factories.job import JobFactory, SubmissionFactory
+from datetime import datetime, timedelta
+from unittest.mock import Mock
+
+from dataactcore.models.jobModels import JobStatus, JobType, FileType
+from dataactvalidator.validation_handlers import file_generation_handler
+from dataactvalidator.validation_handlers.file_generation_manager import FileGenerationManager
+
+from tests.unit.dataactcore.factories.job import JobFactory, FileRequestFactory
 
 
-def test_job_context_success(database, job_constants):
-    """When a job successfully runs, it should be marked as "finished" """
+def test_generate_new_d1_file_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """ Testing that a new D1 file is generated """
     sess = database.session
     job = JobFactory(
-        job_status=sess.query(JobStatus).filter_by(name='running').one(),
-        job_type=sess.query(JobType).filter_by(name='validation').one(),
-        file_type=sess.query(FileType).filter_by(name='sub_award').one(),
+        job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
     )
     sess.add(job)
     sess.commit()
 
-    with file_generation_manager.job_context(job.job_id, is_local=True):
-        pass    # i.e. be successful
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, job)))
+    FileGenerationManager().generate_from_job(job.job_id, '123')
 
     sess.refresh(job)
-    assert job.job_status.name == 'finished'
+    assert job.original_filename != 'original'
+    assert job.from_cached is False
+    assert job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
 
 
-def test_job_context_fail(database, job_constants):
-    """When a job raises an exception and has no retries left, it should be marked as failed"""
+def test_generate_new_d2_file_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """ Testing that a new D2 file is generated """
     sess = database.session
-    job = JobFactory(
-        job_status=sess.query(JobStatus).filter_by(name='running').one(),
-        job_type=sess.query(JobType).filter_by(name='validation').one(),
-        file_type=sess.query(FileType).filter_by(name='sub_award').one(),
-        error_message=None,
-    )
-    sess.add(job)
-    sess.commit()
-
-    with file_generation_manager.job_context(job.job_id, is_local=True):
-        raise Exception('This failed!')
-
-    sess.refresh(job)
-    assert job.job_status.name == 'failed'
-    assert job.error_message == 'This failed!'
-
-
-def test_check_detached_d_file_generation(database, job_constants):
-    """Job statuses should return the correct status and error message to the user"""
-    sess = database.session
-
-    # Detached D2 generation waiting to be picked up by the Validator
     job = JobFactory(
         job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
         job_type=sess.query(JobType).filter_by(name='file_upload').one(),
         file_type=sess.query(FileType).filter_by(name='award').one(),
-        error_message='',
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
     )
     sess.add(job)
     sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'waiting'
 
-    # Detached D2 generation running in the Validator
-    job.job_status = sess.query(JobStatus).filter_by(name='running').one()
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'waiting'
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, job)))
+    FileGenerationManager().generate_from_job(job.job_id, '123')
 
-    # Detached D2 generation completed by the Validator
-    job.job_status = sess.query(JobStatus).filter_by(name='finished').one()
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'finished'
-    assert response_dict['message'] == ''
-
-    # Detached D2 generation with an unknown error
-    job.job_status = sess.query(JobStatus).filter_by(name='failed').one()
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'Upload job failed without error message'
-
-    # Detached D2 generation with a known error
-    job.error_message = 'Detached D2 upload error message'
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'Detached D2 upload error message'
+    sess.refresh(job)
+    assert job.original_filename != 'original'
+    assert job.from_cached is False
+    assert job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
 
 
-def test_check_submission_d_file_generation(database, job_constants):
-    """Job statuses should return the correct status and error message to the user"""
+def test_regenerate_same_d1_file_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """Testing that a new file is not generated if this job already has a successfully generated file"""
     sess = database.session
-    sub = SubmissionFactory()
-    sess.add(sub)
-
-    # D1 generation waiting to be picked up by the Validator
     job = JobFactory(
         job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
         job_type=sess.query(JobType).filter_by(name='file_upload').one(),
         file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
-        submission=sub, error_message=''
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
     )
-    val_job = JobFactory(
+    sess.add(job)
+    sess.commit()
+
+    file_request = FileRequestFactory(
+        job=job, is_cached_file=True, agency_code='123', start_date='01/01/2017', end_date='01/31/2017',
+        file_type='d1', request_date=datetime.now().date(),)
+    sess.add(file_request)
+    sess.commit()
+
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, job)))
+    FileGenerationManager().generate_from_job(job.job_id, '123')
+
+    sess.refresh(job)
+    assert job.original_filename == 'original'
+    assert job.from_cached is False
+    assert job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
+
+
+def test_regenerate_same_d2_file_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """Testing that a new file is not generated if this job already has a successfully generated file"""
+    sess = database.session
+    job = JobFactory(
         job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
-        job_type=sess.query(JobType).filter_by(name='csv_record_validation').one(),
-        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
-        submission=sub, error_message='', number_of_errors=0,
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award').one(),
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
     )
-    sess.add_all([job, val_job])
+    sess.add(job)
     sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'waiting'
 
-    # D1 generation running in the Validator
-    job.job_status = sess.query(JobStatus).filter_by(name='running').one()
+    file_request = FileRequestFactory(
+        job=job, is_cached_file=True, agency_code='123', start_date='01/01/2017', end_date='01/31/2017',
+        file_type='d2', request_date=datetime.now().date(),)
+    sess.add(file_request)
     sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'waiting'
 
-    # D1 generation with an unknown error
-    job.job_status = sess.query(JobStatus).filter_by(name='failed').one()
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'Upload job failed without error message'
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, job)))
+    FileGenerationManager().generate_from_job(job.job_id, '123')
 
-    # D1 generation with a known error
-    job.error_message = 'D1 upload error message'
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'D1 upload error message'
+    sess.refresh(job)
+    assert job.original_filename == 'original'
+    assert job.from_cached is False
+    assert job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
 
-    # D1 generation completed by the Validator; validation waiting to be picked up
-    job.error_message = ''
-    job.job_status = sess.query(JobStatus).filter_by(name='finished').one()
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'waiting'
 
-    # D1 generation completed; validation running in the Validator
-    val_job.job_status = sess.query(JobStatus).filter_by(name='running').one()
+def test_regenerate_new_d1_file_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """Testing that a new file is not generated if another job has already has a successfully generated file"""
+    sess = database.session
+    original_job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='finished').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
+    )
+    sess.add(original_job)
     sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'waiting'
 
-    # D1 generation completed; validation completed by the Validator
-    val_job.job_status = sess.query(JobStatus).filter_by(name='finished').one()
+    file_request = FileRequestFactory(
+        job=original_job, is_cached_file=True, agency_code='123', start_date='01/01/2017', end_date='01/31/2017',
+        file_type='D1', request_date=datetime.now().date(),)
+    new_job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
+        start_date='01/01/2017', end_date='01/31/2017',
+    )
+    sess.add_all([file_request, new_job])
     sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'finished'
 
-    # D1 generation completed; validation completed by the Validator
-    val_job.number_of_errors = 10
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'Validation completed but row-level errors were found'
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, new_job)))
+    FileGenerationManager().generate_from_job(new_job.job_id, '123')
 
-    # D1 generation completed; validation with an unknown error
-    job.error_message = ''
-    val_job.error_message = ''
-    val_job.job_status = sess.query(JobStatus).filter_by(name='failed').one()
-    val_job.number_of_errors = 0
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'Validation job had an internal error'
+    sess.refresh(new_job)
+    assert new_job.original_filename == 'original'
+    assert new_job.from_cached is True
+    assert new_job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
 
-    # D1 generation completed; validation with a known error
-    job.error_message = ''
-    val_job.error_message = ''
-    val_job.error_message = 'D1 upload error message'
-    sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'D1 upload error message'
 
-    # D1 generation completed; validation with an unknown error
-    job.error_message = ''
-    val_job.error_message = ''
-    val_job.job_status = sess.query(JobStatus).filter_by(name='invalid').one()
+def test_regenerate_new_d2_file_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """Testing that a new file is not generated if another job has already has a successfully generated file"""
+    sess = database.session
+    original_job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='finished').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award').one(),
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
+    )
+    sess.add(original_job)
     sess.commit()
-    response_dict = file_generation_manager.check_file_generation(job.job_id)
-    assert response_dict['status'] == 'failed'
-    assert response_dict['message'] == 'Generated file had file-level errors'
+
+    file_request = FileRequestFactory(
+        job=original_job, is_cached_file=True, agency_code='123', start_date='01/01/2017', end_date='01/31/2017',
+        file_type='D2', request_date=datetime.now().date(),)
+    new_job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award').one(),
+        start_date='01/01/2017', end_date='01/31/2017',
+    )
+    sess.add_all([file_request, new_job])
+    sess.commit()
+
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, new_job)))
+    FileGenerationManager().generate_from_job(new_job.job_id, '123')
+
+    sess.refresh(new_job)
+    assert new_job.original_filename == 'original'
+    assert new_job.from_cached is True
+    assert new_job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
+
+
+def test_uncache_same_d1_file_fpds_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """Testing that a new file is not generated if this job already has a successfully generated file"""
+    sess = database.session
+    job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
+    )
+    sess.add(job)
+    sess.commit()
+
+    file_request = FileRequestFactory(
+        job=job, is_cached_file=True, agency_code='123', start_date='01/01/2017', end_date='01/31/2017',
+        file_type='d1', request_date=(datetime.now().date() - timedelta(1)),)
+    sess.add(file_request)
+    sess.commit()
+
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, job)))
+    FileGenerationManager().generate_from_job(job.job_id, '123')
+
+    sess.refresh(job)
+    assert job.original_filename != 'original'
+    assert job.from_cached is False
+    assert job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
+
+
+def test_uncache_new_d1_file_fpds_success(monkeypatch, mock_broker_config_paths, database, job_constants):
+    """Testing that a new file is not generated if another job has already has a successfully generated file"""
+    sess = database.session
+    original_job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='finished').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
+        filename=str(mock_broker_config_paths['d_file_storage_path'].join('original')),
+        start_date='01/01/2017', end_date='01/31/2017', original_filename='original', from_cached=True,
+    )
+    sess.add(original_job)
+    sess.commit()
+
+    file_request = FileRequestFactory(
+        job=original_job, is_cached_file=True, agency_code='123', start_date='01/01/2017', end_date='01/31/2017',
+        file_type='D1', request_date=(datetime.now().date() - timedelta(1)),)
+    new_job = JobFactory(
+        job_status=sess.query(JobStatus).filter_by(name='waiting').one(),
+        job_type=sess.query(JobType).filter_by(name='file_upload').one(),
+        file_type=sess.query(FileType).filter_by(name='award_procurement').one(),
+        start_date='01/01/2017', end_date='01/31/2017',
+    )
+    sess.add_all([file_request, new_job])
+    sess.commit()
+
+    monkeypatch.setattr(file_generation_handler, 'retrieve_job_context_data', Mock(return_value=(sess, new_job)))
+    FileGenerationManager().generate_from_job(new_job.job_id, '123')
+
+    sess.refresh(new_job)
+    assert new_job.original_filename != 'original'
+    assert new_job.from_cached is False
+    assert new_job.job_status_id == sess.query(JobStatus).filter_by(name='finished').one().job_status_id
